refactor: abstract method to be used elsewhere
